Split up adapter into multiple files. (#153) (#154)
* Refactor adapter.rs into multiple files.

* Move edge resolvers to their own file.
