+++ src/adapter/tests.rs
@@ -1,0 +1,18 @@
+use anyhow::Context;
+
+#[test]
+fn rustdoc_json_format_version() {
+    let path = "./localdata/test_data/reexport/rustdoc.json";
+    let content = std::fs::read_to_string(path)
+        .with_context(|| format!("Could not load {path} file, did you forget to run ./scripts/regenerate_test_rustdocs.sh ?"))
+        .expect("failed to load rustdoc");
+
+    let expected_version = rustdoc_types::FORMAT_VERSION;
+    let actual_version = crate::test_util::detect_rustdoc_format_version(&content)
+        .expect("unrecognized rustdoc format");
+
+    assert_eq!(
+        expected_version, actual_version,
+        "Expected to find rustdoc v{expected_version} but got v{actual_version} instead.",
+    );
+}
